zip_my_pet :compress my_pet files
This function is made just to compress the *_my_pet.m and *_my_pet.mat
files. We can make it just to compress the 5 files of each species
(commented lines) or a bit more general all my_pet.m and my_pet.mat
files (if there are more).

for more general set up we will use directly the MATLAB zip function.

Make changes if necessary. This may be the last by me for the time :)